+++ src/plugins/preflight.js
@@ -1,5 +1,6 @@
 import fs from 'fs'
 import postcss from 'postcss'
+import packageJson from '../../package.json'
 
 export default function () {
   return function ({ addBase }) {
@@ -7,6 +8,13 @@ export default function () {
       fs.readFileSync(require.resolve('modern-normalize'), 'utf8')
     )
     const preflightStyles = postcss.parse(fs.readFileSync(`${__dirname}/css/preflight.css`, 'utf8'))
-    addBase([...normalizeStyles.nodes, ...preflightStyles.nodes])
+
+    addBase([
+      postcss.comment({
+        text: `! tailwindcss v${packageJson.version} | MIT License | https://tailwindcss.com`,
+      }),
+      ...normalizeStyles.nodes,
+      ...preflightStyles.nodes,
+    ])
   }
 }